+++ src/InfrastructureModels.jl
@@ -3,6 +3,10 @@ isdefined(Base, :__precompile__) && __precompile__()
 module InfrastructureModels
 
 using JuMP
+
+using MathOptInterface
+const MOI = MathOptInterface
+
 using Memento
 
 import Compat: @__MODULE__
@@ -12,10 +16,7 @@ const LOGGER = getlogger(@__MODULE__)
 
 # Register the module level logger at runtime so that folks can access the logger via `getlogger(InfrastructureModels)`
 # NOTE: If this line is not included then the precompiled `Infrastructure.LOGGER` won't be registered at runtime.
-function __init__()
-    Memento.register(LOGGER)
-    Memento.config(LOGGER, "info")
-end
+__init__() = Memento.register(LOGGER)
 
 include("core/data.jl")
 include("core/relaxation_scheme.jl")